requirements.txt
fastapi
pydantic
sqlalchemy
bcrypt
python-jose

# For use .env files as source of enviroment variables
python-dotenv

# For use postgresql database connection
psycopg2
